+++ lib/src/components/button/raised_button.dart
@@ -22,16 +22,17 @@ class ImpaktfullUiRaisedButton extends StatefulWidget {
   });
 
   @override
-  State<ImpaktfullUiRaisedButton> createState() => _ImpaktfullUiRaisedButtonState();
+  State<ImpaktfullUiRaisedButton> createState() =>
+      _ImpaktfullUiRaisedButtonState();
 }
 
-class _ImpaktfullUiRaisedButtonState extends State<ImpaktfullUiRaisedButton> with SingleTickerProviderStateMixin {
-  var _isPressed = false;
-  late final AnimationController _controller;
-  late final Animation<double> _animation;
+class _ImpaktfullUiRaisedButtonState extends State<ImpaktfullUiRaisedButton>
+    with SingleTickerProviderStateMixin {
+  late AnimationController _controller;
+  late Animation<double> _animation;
   Timer? _timer;
 
-  bool get isDown => _isPressed || widget.isLoading;
+  bool get _isTappable => widget.onTap != null;
 
   @override
   void initState() {
@@ -40,12 +41,25 @@ class _ImpaktfullUiRaisedButtonState extends State<ImpaktfullUiRaisedButton> wit
       duration: const Duration(milliseconds: 25),
       vsync: this,
     );
-    _animation = Tween<double>(begin: 0, end: widget.theme.config.elevation).animate(
+    _createAnimation();
+  }
+
+  void _createAnimation() {
+    _animation =
+        Tween<double>(begin: 0, end: widget.theme.config.elevation).animate(
       CurvedAnimation(
         parent: _controller,
         curve: Curves.easeInOut,
       ),
     );
+  }
+
+  @override
+  void didUpdateWidget(ImpaktfullUiRaisedButton oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    if (oldWidget.theme.config.elevation != widget.theme.config.elevation) {
+      _createAnimation();
+    }
   }
 
   @override
@@ -59,27 +73,26 @@ class _ImpaktfullUiRaisedButtonState extends State<ImpaktfullUiRaisedButton> wit
     if (!widget.theme.config.isRaised) return widget.child;
     final color = _getRaisedBackgroundColor(widget.theme);
     if (color == null) return widget.child;
+    if (!_isTappable) return widget.child;
     return MouseRegion(
       cursor: widget.onTap == null ? SystemMouseCursors.basic : widget.cursor,
       child: GestureDetector(
-        onTapDown: (_) {
-          if (widget.isLoading) return;
-          _timer?.cancel();
-          setState(() => _isPressed = true);
-          _controller.forward();
-        },
-        onTapUp: (_) => _cancelTap(),
-        onTapCancel: _cancelTap,
+        onTapDown: _onTapDown,
+        onTapUp: _onTapUp,
+        onTapCancel: _onTapCancel,
         onTap: widget.onTap,
         child: Stack(
           clipBehavior: Clip.none,
+          fit: StackFit.passthrough,
           children: [
             Positioned.fill(
               child: AnimatedBuilder(
                 animation: _animation,
                 builder: (context, child) => Container(
                   decoration: BoxDecoration(
-                    color: _animation.value < widget.theme.config.elevation ? color : null,
+                    color: _animation.value < widget.theme.config.elevation
+                        ? color
+                        : null,
                     borderRadius: widget.theme.dimens.borderRadius,
                   ),
                   child: child,
@@ -108,8 +121,17 @@ class _ImpaktfullUiRaisedButtonState extends State<ImpaktfullUiRaisedButton> wit
     );
   }
 
-  Future<void> _cancelTap() async {
-    setState(() => _isPressed = false);
+  void _onTapDown(TapDownDetails details) {
+    if (!_isTappable) return;
+    if (widget.isLoading) return;
+    _timer?.cancel();
+    _controller.forward();
+  }
+
+  void _onTapUp(TapUpDetails details) => _onTapCancel();
+
+  void _onTapCancel() {
+    if (!_isTappable) return;
     _timer?.cancel();
     _timer = Timer(const Duration(milliseconds: 30), () {
       _controller.reverse();