+++ lib/src/theme/theme_configurator.dart
@@ -9,7 +9,7 @@ ImpaktfullUiTheme get theme {
   if (_testTheme != null) return _testTheme!;
   if (_theme == null) {
     throw Exception(
-        'Impaktfull theme is not set, make sure you call setImpaktfullTheme() before using theme.');
+        'impaktfull theme is not set, make sure you call setImpaktfullTheme() before using theme.');
   }
   return _theme!;
 }
@@ -17,7 +17,7 @@ ImpaktfullUiTheme get theme {
 Locale get locale {
   if (_locale == null) {
     throw Exception(
-        'Impaktfull locale is not set, make sure you call setImpaktfullLocale() before using theme.');
+        'impaktfull locale is not set, make sure you call setImpaktfullLocale() before using theme.');
   }
   return _locale!;
 }